Allow TransformAction to clone options
When the client is used to query different graph
sources the transform action passed to
CreateFromConfig can be used to manipulate the
keys sent. However the changes are applied to the
global options and doesnt reset after the client
has finished its query.

By introducing a behaviour it is possible to
dictatate if the developer wants to be able to
apply the transformation on a clone of the
options.

The default behaviour is that the current behavior
where the options are global.